% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_query.R
\name{analytics_query}
\alias{analytics_query}
\title{Query the FDW Parameters / Constants / 'Analytics' with an R list defining the query.
This form of the query expects you to pass in the json query in the form of an R list.}
\usage{
analytics_query(
  flight_id,
  query_list,
  efoqa_connection = connect_to_efoqa(),
  start_offset = NULL,
  end_offset = NULL,
  coerce_types = TRUE
)
}
\arguments{
\item{flight_id}{The numeric flight record id to target}

\item{query_list}{list format of the query to perform.  See API documentation for details.}

\item{efoqa_connection}{An optional efoqa connection list object for re-use or customization.}

\item{start_offset}{Request data starting at some offset}

\item{end_offset}{Request data ending at some offset}

\item{coerce_types}{Set this to FALSE to get all results back as strings.  This will improve speed a little bit since no metadata queries will be required.}
}
\value{
A list object with the results of the analytics query.
}
\description{
Query the FDW Parameters / Constants / 'Analytics' with an R list defining the query.
This form of the query expects you to pass in the json query in the form of an R list.
}
\examples{
\dontrun{
query_list_example <- list(
   select = list( list( analyticID = analytic_id_1), list( analyticID = analytic_id_2 ) )
   )

analytics_query(3135409, query_list_example)
}
}
