% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_query.R
\name{analytics_query_from_json}
\alias{analytics_query_from_json}
\title{Query the FDW Parameters / Constants / 'Analytics'}
\usage{
analytics_query_from_json(
  flight_id,
  query_json_file,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{flight_id}{The numeric flight record id to target}

\item{query_json_file}{A reference to a json file describing the query to perform.  See API documentation for details.}

\item{efoqa_connection}{An optional efoqa connection list object for re-use or customization.}
}
\value{
A list object with the results of the analytics query.
}
\description{
Query the FDW Parameters / Constants / 'Analytics'
}
\examples{
\dontrun{
analytics_query_from_json(3135409, query_json_file = "./my_query_file.json")
}

}
