% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_query.R
\name{analytics_query_multiflight}
\alias{analytics_query_multiflight}
\title{Execute an analytics query over multiple flight ids}
\usage{
analytics_query_multiflight(
  flight_ids,
  query_list,
  start_offsets = c(),
  end_offsets = c(),
  efoqa_connection = connect_to_efoqa(),
  coerce_types = TRUE
)
}
\arguments{
\item{flight_ids}{a vector of multiple flight ids to run the analytics query over.}

\item{query_list}{R list form of the analytics query to perform.}

\item{start_offsets}{A vector of start offsets that match with the flight ids.}

\item{end_offsets}{A vector of end offsets that match with the flight ids.}

\item{efoqa_connection}{Optional efoqa connection for multi-system or re-use.}

\item{coerce_types}{Set this to FALSE to get all results back as strings.  This will improve speed a little bit since no metadata queries will be required.}
}
\value{
A dataframe containing the results of executing the query on each flight record.
}
\description{
Execute an analytics query over multiple flight ids
}
