% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_query.R
\name{analytics_query_with_unspecified_flight}
\alias{analytics_query_with_unspecified_flight}
\title{Analytics Query with Unspecified Flight}
\usage{
analytics_query_with_unspecified_flight(
  query_list,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{query_list}{An R list representing the query (equivalent to what you would get by doing jsonlite::read_json() on the json query )}

\item{efoqa_connection}{An optional efoqa connection list object for re-use or customization.}
}
\value{
A list object with the results of the analytics query by selecting a random flight to run the query on.
}
\description{
Analytics Query with Unspecified Flight
}
\examples{
\dontrun{
query_list_example <- list(
   select = list( list( analyticID = analytic_id_1), list( analyticID = analytic_id_2 ) )
   )

analytics_query_with_unspecified_flight( query_list_example )
}
}
