% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apm_profile.R
\name{apm_events_glossary}
\alias{apm_events_glossary}
\title{APM Events Glossary}
\usage{
apm_events_glossary(profile_id, efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{profile_id}{guid string for the profile}

\item{efoqa_connection}{optional efoqa_connection list}
}
\value{
dataframe with an entry for each event type in the profile
}
\description{
APM Events Glossary
}
\examples{
\dontrun{
apm_events_glossary( profile_id = "a7483c44-9db9-4a44-9eb5-f67681ee52b0")
}

}
