% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apm_profile.R
\name{apm_profile_glossary}
\alias{apm_profile_glossary}
\title{APM Profile Glossary}
\usage{
apm_profile_glossary(
  profile_id,
  efoqa_connection = connect_to_efoqa(),
  glossary_format = "json"
)
}
\arguments{
\item{profile_id}{guid string for the profile}

\item{efoqa_connection}{optional efoqa_connection list}

\item{glossary_format}{optional format (csv or json)}
}
\value{
Glossary object representing the APM profile (format 'csv' returns a dataframe, 'json' returns a list)
}
\description{
APM Profile Glossary
}
\examples{
\dontrun{
apm_profile_glossary( profile_id = "a7483c44-9db9-4a44-9eb5-f67681ee52b0")
apm_profile_glossary( profile_id = "a7483c44-9db9-4a44-9eb5-f67681ee52b0", glossary_format = "csv" )
}

}
