% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connect_to_efoqa}
\alias{connect_to_efoqa}
\title{Connect to EMS and get the Auth token.}
\usage{
connect_to_efoqa(
  efoqa_user = Sys.getenv("EFOQAUSER"),
  efoqa_password = Sys.getenv("EFOQAPASSWORD"),
  efoqa_server_url = Sys.getenv("EFOQASERVER")
)
}
\arguments{
\item{efoqa_user}{String, eFOQA username}

\item{efoqa_password}{String, eFOQA password}

\item{efoqa_server_url}{String, eFOQA server}
}
\value{
a Connection object.
}
\description{
Connect to EMS and get the Auth token.
}
