% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_query.R
\name{database_query_from_json}
\alias{database_query_from_json}
\title{database_query_from_json}
\usage{
database_query_from_json(
  data_source_id,
  json_file,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{data_source_id}{String for EMS data source / database ID.  For example FDW Flights = "[ems-core][entity-type][foqa-flights]"}

\item{json_file}{The path to the query you want to run in json file form.  For example "./test_query.json"}

\item{efoqa_connection}{Optional if you want to re-use an existing connection to the API.}
}
\value{
The results of the EMS database query as a dataframe.  Note that no type conversions are performed.  Result types will depend on the 'format' flag in the query.
}
\description{
Executes an EMS database query for the supplied json query file.
}
\examples{
\dontrun{
database_query_from_json("[ems-core][entity-type][foqa-flights]",
  json_file = "./my_query_file.json")
}


}
