% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_query.R
\name{database_query_from_list}
\alias{database_query_from_list}
\title{Execute Database Query from R List}
\usage{
database_query_from_list(
  data_source_id,
  query_list,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{data_source_id}{String for EMS data source / database ID.  For example FDW Flights = "[ems-core][entity-type][foqa-flights]"}

\item{query_list}{R list object containing the EMS API query to execute}

\item{efoqa_connection}{Optional if you want to re-use an existing connection to the API.}
}
\value{
The results of the EMS database query as a dataframe.  Note that no type conversions are performed.  Result types will depend on the 'format' flag in the query.
}
\description{
Execute Database Query from R List
}
\examples{
\dontrun{
database_query_from_json("[ems-core][entity-type][foqa-flights]",
  query_list = standard_flight_query_json)
}

}
