% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_modify.R
\name{delete_from_query}
\alias{delete_from_query}
\title{Delete Entities from Query}
\usage{
delete_from_query(
  data_source_id,
  query_list,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{data_source_id}{Schema moniker of database entity type to be deleted.}

\item{query_list}{List form of the query to get the records that should be deleted}

\item{efoqa_connection}{Connection to efoqa for re-use or advanced use.}
}
\value{
TRUE when completed.
}
\description{
Delete Entities from Query
}
