% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_on_dimensions.R
\name{filter_dimension_by_string}
\alias{filter_dimension_by_string}
\title{Add an additional filter to any query list object that will filter for a specific string value of the target field}
\usage{
filter_dimension_by_string(
  efoqa_connection = connect_to_efoqa(),
  data_source_id,
  query_list,
  field_id,
  dimension_string
)
}
\arguments{
\item{efoqa_connection}{optional existing efoqa connection for re-use or advanced use}

\item{data_source_id}{Schema string for the target entity type (flights or events)}

\item{query_list}{An existing query in R list form}

\item{field_id}{The target field to filter on}

\item{dimension_string}{The target string name of the discrete to filter to}
}
\value{
The input query list with an additional added filter for the target dimension string
}
\description{
Add an additional filter to any query list object that will filter for a specific string value of the target field
}
\examples{
\dontrun{
filter_dimension_by_string( efoqa_connection,
data_source_id = "[ems-core][entity-type][foqa-flights]",
query_list = standard_flight_query_json,
field_id = "[-hub-][field][[[ems-core][entity-type][foqa-flights]][[airframe-engine-field-set][base-field][airframe]]]",
dimension_string = "717-200")
}
}
