% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_on_dimensions.R
\name{get_airframe_id_table}
\alias{get_airframe_id_table}
\title{Get a table of airframes with local ids for this system}
\usage{
get_airframe_id_table(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{optional existing efoqa connection for re-use or advanced use}
}
\value{
A dataframe containing a list of all airframes and their local ID.
}
\description{
Get a table of airframes with local ids for this system
}
\examples{
\dontrun{
get_airframe_id_table()
}
}
