% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_fields.R
\name{get_all_field_details_below_group_as_df}
\alias{get_all_field_details_below_group_as_df}
\title{Get All Field Details for this group and all sub-groups}
\usage{
get_all_field_details_below_group_as_df(
  field_group_id,
  data_source_id,
  exclude_folder_ids = NULL,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{field_group_id}{Target field group EMS schema ID}

\item{data_source_id}{Target entity database or datasource ID}

\item{exclude_folder_ids}{Optional vector of field groups to exclude.  Useful to exclude processing folders.}

\item{efoqa_connection}{optional efoqa connection to use for this operation}
}
\value{
A dataframe with details of all fields in the target field_group_id and all sub-groups.
}
\description{
Get All Field Details for this group and all sub-groups
}
\examples{
\dontrun{
odw_fields <- get_all_field_details_below_group_as_df(
  field_group_id = "[-hub-][field-group][[[odw-logical][entity-type][odw-flights]][[--][internal-field-group][root]]]",
  data_source_id = "[odw-logical][entity-type][odw-flights]",
  exclude_folder_ids = c( "[-hub-][field-group][[[odw-logical][entity-type][odw-flights]][[odw-logical][internal-field-group][flight-info]]]" ) )
}
}
