% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_parameters.R
\name{get_all_physical_parameters}
\alias{get_all_physical_parameters}
\title{Get a list of all physical parameters for all PDCs/LFLs on the system.}
\usage{
get_all_physical_parameters(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{An optional connection for re-use or advanced multi-system use.}
}
\value{
A dataframe with details on all physical parameters (names, units, uids) for all PDCs/LFLs on the system.
}
\description{
Get a list of all physical parameters for all PDCs/LFLs on the system.
}
\examples{
\dontrun{

get_all_physical_parameters()
}
}
