% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_on_dimensions.R
\name{get_database_field_details}
\alias{get_database_field_details}
\title{Get details on any EMS database schema field}
\usage{
get_database_field_details(
  efoqa_connection = connect_to_efoqa(),
  data_source_id,
  field_id
)
}
\arguments{
\item{efoqa_connection}{optional existing efoqa connection for re-use or advanced use}

\item{data_source_id}{The schema name of the data source id that has the target field}

\item{field_id}{The schema name of the target field to get details on}
}
\value{
An R list with details on the target field
}
\description{
Get details on any EMS database schema field
}
\examples{
\dontrun{
get_database_field_details(
data_source_id = "[ems-core][entity-type][foqa-flights]",
field_id = "[-hub-][field][[[ems-core][entity-type][foqa-flights]][[airframe-engine-field-set][base-field][airframe-engine]]]")
}
}
