% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_on_dimensions.R
\name{get_dimension_table}
\alias{get_dimension_table}
\title{Get a table of the avilable discrete states for any field}
\usage{
get_dimension_table(
  efoqa_connection = connect_to_efoqa(),
  data_source_id,
  field_id
)
}
\arguments{
\item{efoqa_connection}{optional existing efoqa connection for re-use or advanced use}

\item{data_source_id}{Schema string for the target data source id}

\item{field_id}{Schema string for the target field id}
}
\value{
A table of the available discrete states for this field along with their local integer id
}
\description{
Get a table of the avilable discrete states for any field
}
\examples{
\dontrun{
get_dimension_table(
efoqa_connection,
data_source_id = "[ems-core][entity-type][foqa-flights]",
field_id = "[-hub-][field][[[ems-core][entity-type][foqa-flights]][[airframe-engine-field-set][base-field][airframe]]]"
)
}
}
