% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_parameters.R
\name{get_logical_FDW_item_list}
\alias{get_logical_FDW_item_list}
\title{Get a list of all FDW logical items ( parameters and constants ) on the system}
\usage{
get_logical_FDW_item_list(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{An optional connection for re-use or advanced multi-system use.}
}
\value{
A dataframe with details on all logical items (names, units, uids)
}
\description{
Get a list of all FDW logical items ( parameters and constants ) on the system
}
\examples{
\dontrun{

get_logical_FDW_item_list()
}
}
