% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odw_fragments.R
\name{get_odw_fragment}
\alias{get_odw_fragment}
\title{Retrieve an ODW Fragment from the EFOQA System}
\usage{
get_odw_fragment(
  efoqa_connection = connect_to_efoqa(),
  fragment_id,
  fragment_format = "file"
)
}
\arguments{
\item{efoqa_connection}{A connection object to the EFOQA system. Defaults to the result of \code{connect_to_efoqa()}.}

\item{fragment_id}{The numeric ID of the fragment to retrieve.}

\item{fragment_format}{A string specifying the format of the fragment. Options are \code{"file"} (default) or \code{"raw"}}
}
\value{
The requested ODW fragment, as a text string.
}
\description{
Retrieve an ODW Fragment from the EFOQA System
}
\examples{
\dontrun{
# Retrieve a fragment in file format
fragment <- get_odw_fragment(fragment_id = "12345")

# Retrieve a fragment in raw format
fragment <- get_odw_fragment(fragment_id = "12345", fragment_format = "raw")
}
}
