% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_fields.R
\name{get_odw_logical_item_fields}
\alias{get_odw_logical_item_fields}
\title{Get all ODW Logical Item fields on this system}
\usage{
get_odw_logical_item_fields(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{optional efoqa connection}
}
\value{
a dataframe with details of the ODW logical item fields
}
\description{
Get all ODW Logical Item fields on this system
}
\examples{
\dontrun{
all_odw_logical_items <- get_odw_logical_item_fields()
}

}
