% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_parameters.R
\name{get_physical_parameters_for_flight}
\alias{get_physical_parameters_for_flight}
\title{Get a list of all physical parameters for a given flight record.}
\usage{
get_physical_parameters_for_flight(
  flight_id,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{flight_id}{The numeric flight id for the flight to get physical parameters for.}

\item{efoqa_connection}{An optional connection for re-use or advanced multi-system use.}
}
\value{
A dataframe with details on all physical parameters (names, units, uids)
}
\description{
Get a list of all physical parameters for a given flight record.
}
\examples{
\dontrun{

get_physical_parameters_for_flight(3135409)
}
}
