% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apm_profile.R
\name{get_profile_group}
\alias{get_profile_group}
\title{get_profile_group}
\usage{
get_profile_group(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{optional efoqa_connection list}
}
\value{
R list describing the profile group contents.
}
\description{
get_profile_group
}
\examples{
\dontrun{
profile_group_content <- get_profile_group( )
}
}
