% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_modify.R
\name{insert_data_frame}
\alias{insert_data_frame}
\title{Insert Entities from Data Frame}
\usage{
insert_data_frame(
  input_df,
  schema_map,
  data_source_id,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{input_df}{An R dataframe that should be inserted into the EMS database}

\item{schema_map}{An R list that maps the dataframe columns to schema monikers.}

\item{data_source_id}{Schema moniker of database entity type to be deleted.}

\item{efoqa_connection}{Connection to efoqa for re-use or advanced use.}
}
\description{
Insert Entities from Data Frame
}
