% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apm_profile.R
\name{list_all_apm_profiles}
\alias{list_all_apm_profiles}
\title{list_all_apm_profiles}
\usage{
list_all_apm_profiles(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{optional efoqa_connection list}
}
\value{
A list off all profiles on the system, including names, ids, tree locations, etc.
}
\description{
list_all_apm_profiles
}
\examples{
\dontrun{
all_profiles <- list_all_apm_profiles( )
}
}
