% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_query.R
\name{search_for_analytic}
\alias{search_for_analytic}
\title{Search for Analytic IDs}
\usage{
search_for_analytic(text_to_search, efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{text_to_search}{Character string to search for in the list of 'analytics'}

\item{efoqa_connection}{Optional connection list object for re-use and customization}
}
\value{
list containing all analytics found with the search text
}
\description{
Search for Analytic IDs
}
\examples{
\dontrun{
search_for_analytic("Maximum Operating Altitude")
}

}
