% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_queries.R
\name{standard_airframe_query}
\alias{standard_airframe_query}
\title{Standard Airframe Query}
\usage{
standard_airframe_query(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{Optional connection string for re-used or advanced configuration}
}
\value{
A dataframe with one entry per airframe-engine on the system giving the number of flights, and most recent dates and flight records
}
\description{
Standard Airframe Query
}
