% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_queries.R
\name{standard_event_query}
\alias{standard_event_query}
\title{Standard All Events Query}
\usage{
standard_event_query(
  event_data_source_id =
    "[ems-apm][entity-type][events:profile-a7483c449db94a449eb5f67681ee52b0]",
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{event_data_source_id}{String representing the datasource ID for events in an EMS profile.}

\item{efoqa_connection}{Connection list for re-use.}
}
\value{
A dataframe with the standard event fields
}
\description{
Executs a query for all events in the target profile and returns the standard global event fields.
This function defaults to running on the library flight safety events profile, but can be modified to use other profiles by changing the data_source_id
}
\examples{
\dontrun{
standard_event_query()
ds_id = "[ems-apm][entity-type][events:profile-87d17d3479804596a231d1d875c85e1f]"
standard_event_query( ds_id )
}
}
