% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_queries.R
\name{standard_fleet_query}
\alias{standard_fleet_query}
\title{Standard Fleet Query}
\usage{
standard_fleet_query(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{Optional connection string for re-used or advanced configuration}
}
\value{
A dataframe with one entry per fleet on the system giving the number of flights, and most recent dates and flight records
}
\description{
Standard Fleet Query
}
