% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_queries.R
\name{standard_flight_query}
\alias{standard_flight_query}
\title{standard_flight_query}
\usage{
standard_flight_query(efoqa_connection = connect_to_efoqa())
}
\arguments{
\item{efoqa_connection}{optionally supply an existing connection to the EMS API for re-use.}
}
\value{
A dataframe containing one row per FDW flight with valid data
and columns with some of the most commonly needed fields.
}
\description{
Executes a query for FDW flights returning some of the most commonly used fields.
}
\examples{
\dontrun{
standard_flight_query()
}
}
