% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{trusted_connect_to_efoqa}
\alias{trusted_connect_to_efoqa}
\title{Connect to EMS and get the Auth token using Trusted Authentication.}
\usage{
trusted_connect_to_efoqa(
  efoqa_client_id = Sys.getenv("EFOQACLIENTID"),
  efoqa_client_secret = Sys.getenv("EFOQACLIENTSECRET"),
  trusted_auth_name = Sys.getenv("EFOQATRUSTEDAUTHNAME"),
  trusted_auth_value = Sys.getenv("EFOQATRUSTEDAUTHVALUE"),
  efoqa_server_url = Sys.getenv("EFOQASERVER")
)
}
\arguments{
\item{efoqa_client_id}{String, eFOQA client ID to use for trusted authentication}

\item{efoqa_client_secret}{String, eFOQA client secret to use for trusted authentication}

\item{trusted_auth_name}{String, eFOQA property to search for as part of trusted authentication}

\item{trusted_auth_value}{String, value of the eFOQA property to use for trusted authentication}

\item{efoqa_server_url}{String, eFOQA server}
}
\value{
a Connection object.
}
\description{
Connect to EMS and get the Auth token using Trusted Authentication.
}
