% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_query.R
\name{analytics_query_multiflight}
\alias{analytics_query_multiflight}
\title{Execute an analytics query over multiple flight ids}
\usage{
analytics_query_multiflight(
  flight_ids,
  query_list,
  efoqa_connection = connect_to_efoqa()
)
}
\arguments{
\item{flight_ids}{a vector of multiple flight ids to run the analytics query over.}

\item{query_list}{R list form of the analytics query to peform}

\item{efoqa_connection}{optional efoqa connection for multi-system or re-use.}
}
\value{
A dataframe containing the results of executing the query on each flight record.
}
\description{
Execute an analytics query over multiple flight ids
}
